refactor: moved all types to types folder
