+++ src/types/index.ts
@@ -11,3 +11,5 @@
  export * from './transaction';
  export * from './utils';
  export * from './wallet';
+ export * from './db';
+ export * from './machine';
