+++ src/types/machine.ts
